XMonad: Shift focus and windows with arrow keys
I had to get used to the idea that XMonad sees windows as a
left-to-right, top-to-bottom list instead of a spatial layout like i3,
but I like it more, because with i3 I was never sure where I would end
up if I moved the focus between columns. Now Arrows keys make everything
more sensible.