NVim: Add airline status line
I wanted to display the current mode and file type in the status bar.
Everything else is a nice bonus.

In the future, I'd like to avoid the "jumping and blinking" caused by the
mode display in section A. This could be done by replace the mode
display by 2 characters.