BC break: Manage more apps with stow
Move directories for better compatibility with GNU `stow`.
After re-reading some tutorials I realized that I can manage the
`.config` directory with `stow` as well, by creating a file structure.

This is slightly inconvenient in fish autocompletion and NVim Telescope,
because the `.config` file inside the folders hides the files, but I
think the `stow` compatibility is worth it.

This is a BC break of existing installations, you need to delete your
symlinks and re-run stow!