NVim: Add GitHub Copilot
Remove Tab behavior form nvim-cmp to make it work. I think that's fine
because I don't remember using tab anyway.

I chose to use the official github/copilot because it was the easiest
and I trust Tim Pope.

I've seen bug reports of vim-cmp-copilot, to which I might switch in the
future