fish: Delete oh-my-fish config loader
I no longer use much of oh-my-fish, since I'm no longer
using "marlin" path completion but use zoxide instead. The last
remaining bit is the `bobthefish` theme, but that does not work well
with Jujutsu VCS, so even that might be replaced.

You can still place the omf init in `fish/conf.d`, because that
directory is in `.gitignore` anyway, because it's supposed to contain
system-specific initializations.